+++ pidx/core/PIDX_header/PIDX_header_io.c
@@ -918,7 +918,7 @@ static int populate_meta_data(PIDX_header_io_id header_io_id, PIDX_block_layout 
   off_t data_offset = 0, base_offset = 0;
   //int total_file_size = 0;
 
-  int64_t total_chunk_size = (header_io_id->idx->chunk_size[0] * header_io_id->idx->chunk_size[1] * header_io_id->idx->chunk_size[2] * header_io_id->idx->chunk_size[3] * header_io_id->idx->chunk_size[4]);// / (64 / header_io_id->idx->compression_bit_rate);
+  unsigned long long total_chunk_size = (header_io_id->idx->chunk_size[0] * header_io_id->idx->chunk_size[1] * header_io_id->idx->chunk_size[2] * header_io_id->idx->chunk_size[3] * header_io_id->idx->chunk_size[4]);// / (64 / header_io_id->idx->compression_bit_rate);
 
 
   /*
